requirements.txt
Django==5.1.7
djangorestframework==3.15.2
drf-spectacular==0.28.0
coverage==7.7.1

sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
asgiref==3.8.1
whitenoise==6.9.0
requests==2.32.3